Made connector lines place accurately
Connector lines previously were placed from top to bottom, not
accounting for line thickness. Therefore, at the top, they would
be right at the top, and near the bottom, they would get closer
and closer to the bottom. Basically, they were not placed on the
same scale that lines were. I added line thickness / 2 to either
side, so that they are placed on the correct scale.